requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
astroid==3.0.3
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==24.1.1
bleach==6.1.0
blinker==1.9.0
celery==5.3.4
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.1
charset-normalizer==3.4.4
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
coverage==7.13.5
cron_descriptor==2.0.8
cryptography==42.0.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
Django==4.2.11
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.1
django-mysql==4.12.0
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-csv==3.0.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.0
et_xmlfile==2.0.0
executing==2.2.1
factory-boy==3.3.0
Faker==22.2.0
fastjsonschema==2.21.2
flake8==7.0.0
Flask==3.1.3
Flask-Login==0.6.3
fqdn==1.5.1
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kombu==5.6.2
lark==1.3.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.2.0
multidict==6.7.1
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
mysqlclient==2.2.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.4
notebook_shim==0.2.4
numpy==2.4.4
opencv-python==4.13.0.92
openpyxl==3.1.2
overrides==7.7.0
packaging==26.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
Pillow==10.1.0
platformdirs==4.9.2
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==3.0
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.12.1
pylint==3.0.3
PyMySQL==1.1.1
pyotp==2.9.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2026.1.post1
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.1
referencing==0.37.0
reportlab==4.0.9
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
Send2Trash==2.1.0
sentry-sdk==1.40.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tomlkit==0.14.0
tornado==6.5.4
traitlets==5.14.3
twilio==8.11.0
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.8
whitenoise==6.6.0
yarl==1.23.0
